Use dummy data
